implemented adding rating feature

+++ src/main/java/com/unibuc/rolls_dice/entity/BoardGameCollection.java
@@ -24,7 +24,7 @@ public class BoardGameCollection {
     @JoinColumn(name = "board_game")
     private BoardGame boardGame;
 
-    private Integer ratingScore;
+    private Float ratingScore;
     private String ratingDescription;
 
     @Column(nullable = false)
